Couple of changes
Changed the file name for the html in studio 2, to index.html. (That is what it will need to be when we turn it in). Also changed the css file back to css extension and just commented out the block of code that makes the red boxes that way it is easier to change back.