A whole lotta change....jk
Copied the first box you made and created a new field set below it and changed the selection type from radio to checkbox. Also set the default "checked" ones for both boxes. Created the paragraph below the boxes along with the listing of advantages and linked them to the correct page. Along with making the pages open in a new tab rather than the tab currently being used.