Lots of changes - Mostly bottom three boxes
In the box with the poem thing. I created a inner shadow mainly as a place holder. Will need to get exact styling, color, and capitalize the L.

Will have to adjust sizing for the bottom three div boxes. I created a inner border for the text and made it like a pink will need to adjust it to make it more red.
In the third box we will need to adjust the legend to make it a rounded border.
Made it to where when highlighting one of the selections in the third box it changes color when using “hover” effect.

Let me know if you have any questions on it