Last box has been colorized. Webpage done.
the webpage should now be perfect! we may need to shrink down the code a little, however certain sections will be impossible to condense. For example, the single-letter selector must be used for the lorem ipsum text, so to bold it, we must have a separate selector for JUST that letter.