Consolidating and less specific
Consolidated code a little bit. FIgured out how to combine the two text boxes with the reddish borders around the text. Made rounded border in the fieldset around choices. Made code less specific, but it still does what it is supposed to do. Not sure if there is anything else we could make shorter.