Made specific comments, started findUniqueLetters and code book
Made specific comments. Started filling some smaller parts on findUniqueLetters. Messed with it quite a bit but couldn't ever get it to work, so moved onto the code book. The code book is almost done. It will store and print the codeword and meaning if you enter both or if you enter codeword and leave the meaning blank. But a couple of the buttons don't seem like they are working and the "meaningless" string never gets printed.